requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pyupbit==0.2.33
pandas==2.2.0
python-dotenv==1.0.1
jinja2==3.1.3
aiohttp==3.9.3
